Remove action logic form season end scheduler and reorganise files
