+++ src/discord/seasonWinnerRoleUpdater/types.ts
@@ -1,5 +1,4 @@
 export interface SeasonWinnerRoleUpdateRequest {
-  action?: string;
   seasonKey?: string;
   source?: string;
 }
